agent-hiring-mvp/agenthub-sdk/templates/acp_agent/requirements.txt
acp-sdk
smolagents
python-dotenv